regexp: fix [Aa] character class bug
This fixes a problem with match.go:stepByte().  If we need to case fold
for this particular instruction, we uppercase the character 'c', but it
doesn't get reset back to the lowercase version when we move on to
processing the next state -- the character 'c' is still the modified
uppercase version instead of the original lowercase version that's
actually in the string.  I've included two unit tests for this fix.

Fixes issue 8.

R=rsc
http://codereview.appspot.com/6052053