index: adjust code in preparation for large indexes
The old code was written with care to avoid problems
with >2GB indexes on 32-bit machines, using uint32
offsets throughout. But then it mmap'ed the index file,
which wouldn't work with a >2GB index anyway, so none
of that code was really useful in the end.

Move the code to use int for offsets and values throughout.
This will still not work well with >2GB indexes on 32-bit
machines, but again the code already doesn't work in that case.
Now it may at least diagnose the problem a bit better.

The old uint32 code did work well for up to 4GB indexes
on 64-bit machines. This is in preparation for >4GB indexes.